snmp__apc: have a static title for the graph.
Instead of adding model and serial number to the graph itself, use a
static title for it and instead provide the model and serial number as
part of graph_info.

This change makes it much easier to forward the warning/critical
statuses back to Icinga (or Nagios), where you have to match the graph
title.