Cleanup APC UPS plugins using SNMP.
Remove the plugins using snmpwalk to produce their results, as there
are three different, Perl-based implementations.

Of these, one might not work because it uses net-snmp and thus rely on
the presence of some MIB file which might not be present.

None of the plugins use Munin::SNMP and none uses multigraph
capabilities.