hwmon: initial import of a sysfs-based plugin to monitor Linux hardware sensors.
This is designed as a replacement for the sensors_ plugin in Munin; it
accesses the same data as that plugin, but does so through direct
access to /sys, following the kernel-mandated interface, instead of
using the `sensors` command.

This allows much faster and reliable monitoring, especially on
dme1737-based systems where the driver is prone to failures.

Compared to the sensors_ plugin, this does not let you change the
warning/critical levels from the node configuration, nor to ignore any
particular sensor, as it provides whatever the kernel is providing.

Also, this plugin will show humidity and current readings, if present.