apache_status: add a new multigraph plugin to replace the old apache_* from the main repository.
This plugin uses a single requests to parse all the statistics instead
of making three of them, reducing the overhead, and uses the
multigraph capabilities to generate the same graphs. The naming is
compatible so that it should be a clean update from the previous
plugins.

This also merges in apache_average_requests (and apache_request_rate)
which, albeit not as relevant, also parsed the same file.