Fixes bandwidth_ to successfully init new installs.
Previously the script would try to divide by 0 for calculating rates
on the first run, and crash before getting to store() to write initial
values thus never initing itself. This fixes that, and makes it a
little easier to debug in case of future crashes (no need for 'eval'
wrapper).