add plugin to monitor "smart meters" power usage
this connects munin with the RTLAMR project, and allows readout of AMR
(Automatic Meter Reporting) data over the airwaves

This plugin will draw some nice graphs of your neighbors (and
yourself, coincidentally) power consumption. You can also see the
adoption of the AMR in your neighborhood and density of signal.