General purpose backup monitoring plugin
If you have a directory where daily, weekly, monthly you generate
backups of files, SQL databases, etc, you might want to be informed
if the backup generation process is gone or the behavior changes
greatly. This monitoring plugin watches the number of young files
in a certain directory. Relies on find command to do the job.