zfs-filesystem-graph: remove --upper-limit
Providing a --upper-limit parameter is only useful if the size of the
filesystem never changes. On ZFS, this assumption is only true for the
zpool; for all other filesystems, the available space is shared, and the
totalled values will vary. When this is the case, the --upper-limit
parameter will obscure peaks in the graph.