Plugin runs, even if no access to list DBs.
If the user as whom "munin" connects has no rights to list existing DBs,
the plugin crashes. Even though the rest of the code is OK. Listing DBs
is only necessary in the case of auto-configuring the plugin. Not for
running it.

This removes the line which causes the bug. The variable which is set in
this line is anyway not used.