new network plugin added: wireless_channel_occupation
This plugin tracks the occupation of the shared medium (wifi channel).
It uses the "iw" tool for extracting the time slices spent by
the wireless driver receiving, transmitting or ignoring (noise) on the
current channel. This ratio is a good indication for the remaining
capacity of the shared communication medium.