[monit_parser] implement memory alloaction unit parsing
Previously monit seems to have output memory usage in kilobytes without
a unit. Somewhen this seemed to have changed to an (optional?) suffix
(e.g. MB). Thus the new scaling may differ from the previous scaling,
which was probably broken for most users.
Additionally the units of the resulting values are clarified.