jenkins: Added support for jobs folders
Newer versions of Jenkins have support for grouping jobs into "folders",
for example when using a multibranch setup, the main "job" will actually
be a folder containing a job for each branch.

This changeset adds a "jobDepth" parameter that can be used to control
how far down the folder tree to report on.