Bugfix: fix graph scaling in munstrap.
Bootstrap sets box-sizing to border-box to make lay-outing easier, but
this damages the scaling of the munin graphs, which get a border to
indicate their warning state.

This patch makes all images have box-sizing: content-box, so that the
graphs are actually the intended size, rather than the intended size
minus the border, causing a blurry image.