tests: work around 'checkbashisms' complaining about "command -v"
"command -v" should be the preferred way of determining the location (or
existence) of an executable.
Sadly "checkbashisms" interprets "command -v" as a possible bashism,
since "command -v" was introduced in IEEE 1003.1 2013.
But checkbashisms checks for compliance with IEEE 1003.1 2004 - here
"command -v" is listed only as a portability extension.

See https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=733511 for the
discussion about checkbashism and "command -v".

See https://unix.stackexchange.com/a/85250 for the discussion about
"command -v" being the most portable check for a program's location.

The workaround itself is bit ugly: 'checkbashisms' is monkey-patched
and fed into the perl interpreter. This makes it harder to re-run the
check command manually (due to complicated quoting).