kvm_net: improve the network interface name parser and admit its limits
The network interface parser of this plugin was overly specific before.
It relied on a specific format of the arguments handed over to kvm while
starting the VM.  For example the following format was usable:

  ... -netdev tap,ifname=foo,...

But kvm/qemu support a variety of ways for configuring network
interfaces via the commandline.  E.g. libvirt does not use the "ifname"
parameter above.  Thus VMs running on a host controlled via libvirt
cannot be tracked with this plugin.

This limititation is now clearly documented in the header of the plugin.