asterisk: Add configuration options to disable conference graphs
Since the MeetMe application is not available anymore on Debian, it
makes no sense to keep the graph around. We do however want to keep code
around for parsing stats from MeetMe. So we'll add two "binary" flags
that'll let users which of these two graphs they want.

The default values will favor getting information from ConfBridge which
is the suggested replacement for MeetMe.

An added bonus of this is that some users might decide to disable both
graphs if they use neither applications, which was impossible before.