tests: enable shellcheck for shell-based plugins and flake8 for python-based plugins
Many plugins are still failing these tests.
These expected failures are listed in a file
(t/test-exception-wrapper.expected-failures)
A wrapper script is used for running the tests and comparing the result
with the expectation (based on the file being listed in the above file).

This allows to test all new plugins, while ignoring all known failures.