Remove obsolete bebox plugins
The plugins beboxsync, beboxstats and beboxstats.expect are obsolete.
The ISP that sold the Bebox [Be*] was acquired by another provider
more than five years ago so it is very unlikely that anybody is using
these plugins any more.

The plugin tg585v7_ is a suitable replacement as the Bebox and TG585
are both made by Technicolor [formerly Thompson] and run on the same
platform. tg585v7_ also seems to be much more comprehensive than the
bebox plugins.

Closes: #1023
Thanks, troffasky