create new plugin for graphing PuppetDB's usage of JVM Heapspace
PuppetDB requires a pretty high amount of memory allocated to it. If you
don't give enough memory space to the JVM, PuppetDB might crash when it
has fully used the available max memory. This plugin should be useful
for better tuning up your instance(s) of PuppetDB.

Currently, it's only compatible with PuppetDB 6.x since it's requesting
data from the API as described for that version:

https://puppet.com/docs/puppetdb/6.0/api/metrics/v1/mbeans.html#jvm-metrics

To add support for more versions, one would need to consult the same
documentation for different versions and to adapt URLs used and the code
that's obtaining the data from JSON if the layout of the responses is
different.