relayd: avoid matching table defs within comments
if a table is commented out, you currently get an entry for the hosts
that the entry contains but you'll get NaN for those hosts all the time
if they are not defined elsewhere.

to avoid that we anchor our match to the beginning of lines and permit
leading spaces since some folks might like to indent lines.