Update the Deluge plugin to handle Deluge 2.x
Version 1.0.0 of this plugin requires version 1 of Deluge. Deluge 2 was
released in June 2019.

This commits bumps the plugin to version 2.0.0.

This plugin has three modes. The modes "bandwidth" and "states" are
working equally to the previous version.  The third mode, "connections",
is changed.

The connections mode previously printed one value: the total number of
connections.  That RPC call is removed in Deluge 2, and I'm instead
using `get_session_status`, which delegates the call to libtorrent [1].
libtorrent is much better documented, so it's easier to work with
compared to Deluge.

"Number of peers" replaces the previous "number of connections", and
I've added half open peers, and peers interested in download and upload
slots to the same graph.

[1]. https://libtorrent.org/manual-ref.html#session-statistics