chrony_status: fix fetch output for serverstats graph
Was not expecting #1255 to be merged so quickly 🤪. Now the new
serverstats graph has been fully tested and actually works 😉.

This change also doesn't needlessly convert all values to floating
point numbers, instead provides the values to munin in the same format
as they are originally output by chrony.

Also this adds some server stats found in chrony 4.0 (but still works
with 3.4 and probably earlier).