Add gpsd_ plugin
This is a multigraph plugin which graphs various stats from gpsd. Only
enough statistics are graphed to show the _quality_ of the fix, not the
actual fix itself (attempting to protect privacy).

GPSd will report various messages asynchronously (that is, the messages
may come in any order) so, during the collect phase, we emit values as
and when the messages arrive. This should provide the most expedient
response.