fix(weather_press_/temp_): Cast url-read object to string
The read function on an urllib urlopen object returns an object as a response.
Regular expressions using re can't be used on such objects. This causes
the following error:

```
Traceback (most recent call last):
  File "/tmp/weather/./weather_press_LOWW", line 43, in <module>
    hpa = re_hpa.findall(txt)[0]
TypeError: cannot use a string pattern on a bytes-like object
```

This can be easily fixed, because said object can simply be cast to string.
Which is, what this patch does for both the US NOAA based plugins.