fix(weather_temp_): Better match temperature and dew point
For whatever reason, sometimes linebreaks are only denoted as \n in the
querry-result. This lead to the temperature RegEx matching the dew point. This
caused the temperature to be reported the same as the dew point. These changes
should make the RegEx more specific and work consistently, no matter if
linebreaks are actual linebreaks or just \n.