Add Hibiscus server homebanking plugins
1. Last account update age in hours
2. Account balances

The plugins have been in use for several months now and warn
when the balance of accounts drop below a configurable value,
and when Hibiscus stopped updating data because a TAN needs to
be entered manually after 90 days.

The plugins directly query the MySQL database of a Hibiscus server
instance.
See https://www.willuhn.de/products/hibiscus-server/ for more information.