gpsd: add C rewrite
The graphs are basically identical to the original Ruby plugin, but
there is no longer a dependency on the "gpspipe" utility; a direct TCP
connection to gpsd is established, and the JSON is parsed using the
json-c library.

The intent here is to provide the same functionality on systems that are
resource-constrained and/or sensitive to jitter (e.g., an embedded NTP
server).

This can be compiled like such after editing GPSD_HOST / GPSD_PORT:

	gcc -o gpsd gpsd.c -ljson-c