rtorrent: make plugins compatible with TinyXML2-generated responses
The plugins were written in a way that depended on the XML having
newlines in specific places. Building rtorrent with TinyXML2 leads to
responses with no newlines, breaking the plugins.

Normalize XML responses to remove whitespace and adjust the parsers
accordingly.

Additionally:
- Deduplicate some of the code (UNIX vs TCP) in the 'allsession'
variants
- Add basic response checking to a couple of the plugins (rtom_mem and
rtom_spdd) so that they return 'U' if the expected values aren't found
- Make several of the regex groups non-capturing where the data isn't
wanted
- Fix broken links to wiki
- Remove references to '--with-xmlrpc-c' build flag
- Remove unused 'max_rate' fetches in rtom_allsession_spdd
- Switch indents to tabs in a few places; this matches the majority of
lines throughout the set of plugins